feat(derive): Implement via Iterator for flatten
Adds the first `via` field attribute initially only for ToJson when used
with flatten. Via will be similar to serde_with, generally some of name of
a trait to be used as a in between.

One obstacle for via = Iterator is determine what exact time we should be
iteratoring over. For instance, for `x: HashSet<T>`, we what iterator over,
&HashSet<T> (`&x`) but for `y: &[T]` we want to iterate over `&[T]` (`y`).
Currently, we just call `.iter()` but in future we should probably switch to
`IntoIterator` and then add refence if not already a reference.