refactor: Seperate parser and scratch buffer
Due the lack of partial borrows in stable rust previously we had to rely
on unsafe to erase the lifetimes of strings that may have been borrowed
from the scratch buffer when we need mutate parser but keep the strings.

Now with the seperation, the borrow checker can see that parser and buffer
have disjoint lifetimes allowing the limited access to the parse without
unsafe.

This did come at a cost about ~1-2% runtime and compile-time performance.

One similar case where unsafety is still needed is when passing the
borrowed key and parser to a function. Rust currently doesn't have a way
to safely represent the lifetime relation in function arguments. This
is required for field visitors and now the unsafety has be encapsulated
with ParserWithBorrowedKey. Although construction of `ParserWithBorrowedKey`
is still unsafe the vistor method is not.