fix(json): Unify lazy parser for MaybeJSON
Previously, the lazy parser for MaybeJSON was it's own implementation
that shared some code for parsing strings. Additionally, the error codes
where unified as well.

The missing object key error storage has been simplified and use cases
restricted. Previously, we would attempt store the length and pointer
to the string in a single pointer and thus relied on platform specific
unused high bits of the pointer.

Instead, we now only attempt to store the pointer when index with an
extra layer of indirection (&'static &'static str). Since then we can
write we use alignment bits for pointer tagging and no longer need to
store the tag.