fix(template): Return String even on compile error
Previously, we would return an block containing a compiler_error!(...),
this works great for simple cases, but since the block would now evaluate
to `()` instead of `String` it could cause a cascade of errors that can
overshadow the root cause error message within the compiler_error! block.

Now, we return something like:
{
    compiler_error!(..)
    ::std::string::String::new()
}
So that now the block continues the evaluating to a String.