feat(attr): Add #[jsony(validate = ...)] field attr
Previously, to add additional requirements to a value, you would have to
define manual `impl`s using a `with` attribute or by defining a newtype.

The `validate` field attribute takes a closure of the form:
`|&T| -> Result<(), String>`. To aid in generating the closure for simple
cases, a `jsony::require` macro was added. In the future, we may extend
validation to support common cases directly. This will be enabled by supporting
the following syntax:

`#[validate = {min: 3, max: 32, message: "Some Message"}]`

That extension to `validate` hasn't been fully explored yet, but we have
left room for a non-breaking change to introduce such functionality.

Let me know if you'd like a more formal tone or documentation-ready version.